% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcsim.r
\name{lines.qcsim}
\alias{lines.qcsim}
\title{Add QC power curve}
\usage{
\method{lines}{qcsim}(x, lwd = 2, ...)
}
\arguments{
\item{x}{object returned from qcsim}

\item{lwd}{line width}

\item{...}{arguments to be passed to methods, such as graphical parameters.  See base::plot() for details.}
}
\description{
Add QC power curve
}
\examples{

iters <- 51 # number of x points
maxs <- 5

pres.bias.3s <- qcsim(maxs=maxs, xsize=iters)
pres.bias.W <- qcsim(maxs=maxs, xsize=iters,
                     rule1_sigS=TRUE,rule2_2S=TRUE,ruleR_4S=TRUE,rule4_1S=TRUE,rule10X=TRUE)

# Overlay for comparison
plot(pres.bias.3s)
lines(pres.bias.W, col=2)
}
