% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcsim.r
\name{plot.qcsim}
\alias{plot.qcsim}
\title{Plot QC power curve}
\usage{
\method{plot}{qcsim}(x, lwd = 2, ...)
}
\arguments{
\item{x}{object returned from qcsim}

\item{lwd}{line width}

\item{...}{arguments to be passed to methods, such as graphical parameters.  See base::plot() for details.}
}
\description{
Plot QC power curve
}
\examples{

iters <- 51 # number of x points
maxs <- 5

# 3 SD rule
pres.bias.3s <- qcsim(maxs=maxs, xsize=iters)
plot(pres.bias.3s, col=1)

}
