% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcsim.r
\name{qcsim}
\alias{qcsim}
\title{Simulate quality control rule performance}
\usage{
qcsim(
  numiter = 2e+05,
  maxs = 5,
  plex = 1,
  levels = 1,
  reps = 1,
  xsize = 51,
  sig = 3,
  num.fails = 1,
  precision = FALSE,
  ruleP_2S = TRUE,
  rule1_sigS = TRUE,
  rule2_2S = FALSE,
  ruleR_4S = FALSE,
  rule4_1S = FALSE,
  rule10X = FALSE,
  ruleT = FALSE,
  cutoff.T = 0,
  ruleF = FALSE,
  cutoff.F = 0,
  FTrefLength = 10,
  FTtestLength = 10,
  byLevel = TRUE,
  cores = 1,
  seed = NULL
)
}
\arguments{
\item{numiter}{number of empirical trials/point}

\item{maxs}{max sigma for calculation (e.g. maxs = 5 gives results from a delta of 0s to 5S)}

\item{plex}{plex of the assay (e.g. 1, 10, 100)}

\item{levels}{# of levels of QC material per time point}

\item{reps}{# of replicates of each QC level per time point}

\item{xsize}{# of discrete points to calculate across the domain}

\item{sig}{significance cutoff for first-level (rule1) QC.  Default is 3S.}

\item{num.fails}{number of failures allowed per multiplex test}

\item{precision}{are we iterating across precision (T) or accuracy/bias (F)}

\item{ruleP_2S}{Westgard preliminary 2s threshold}

\item{rule1_sigS}{Westgard threshold at "sig" SD (default 3S)}

\item{rule2_2S}{Westgard 2_2S rule}

\item{ruleR_4S}{Westgard R_4S rule}

\item{rule4_1S}{Westgard 4_1S rule}

\item{rule10X}{Westgard 10x rule}

\item{ruleT}{"T-test" rule (always byLevel)}

\item{cutoff.T}{cutoff P-value for ruleT}

\item{ruleF}{"F-test" rule (always byLevel)}

\item{cutoff.F}{cutoff P-value for ruleF}

\item{FTrefLength}{reference population length (T and F test)}

\item{FTtestLength}{test population length (T and F test)}

\item{byLevel}{additional lookback via consecutive samples at one level for 2_2S, 4_1S, 10X}

\item{cores}{# of cores for parallel processing}

\item{seed}{seed value for reproducible random sequence}
}
\value{
qcsim object

import(Rcpp)
}
\description{
Simulate quality control rule performance
}
\examples{

iters <- 51 # number of x points
maxs <- 5

# Note: running the function for bias shift returns deltaSE from 0 to maxs,
#   while running for precision shift returns deltaRE from 1 to (maxs + 1)

par(mfrow=c(1,2))

# 3 SD rule
pres.bias.3s <- qcsim(maxs=maxs, xsize=iters)
plot(pres.bias.3s, col=1)
pres.prec.3s <- qcsim(maxs=maxs, xsize=iters, precision=TRUE)
plot(pres.prec.3s, col=1)

# Full Westgard rules
pres.bias.W <- qcsim(maxs=maxs, xsize=iters,
                     rule1_sigS=TRUE,rule2_2S=TRUE,ruleR_4S=TRUE,rule4_1S=TRUE,rule10X=TRUE)
plot(pres.bias.W)
pres.prec.W <- qcsim(maxs=maxs, xsize=iters, precision=TRUE,
                     rule1_sigS=TRUE,rule2_2S=TRUE,ruleR_4S=TRUE,rule4_1S=TRUE,rule10X=TRUE)
plot(pres.prec.W)


# Overlay for comparison
plot(pres.bias.3s)
lines(pres.bias.W, col=2)
}
